use uint64_t instead of uint32_t in for loops with condition <=
Using uint32_t as index with <= condition may result in overflows when
right side of condition is UINT32_MAX. In such case when the index has
value of UINT32_MAX the condition is stil true, the loop will execute,
and the result of UINT32_MAX + 1 is 0 (overflow) and the condition will
be still true - that results in infinite loop.

+++ src/herder/HerderImpl.cpp
@@ -1031,7 +1031,8 @@ HerderImpl::sendSCPStateToPeer(uint32 ledgerSeq, PeerPtr peer)
         minSeq = maxSeq = ledgerSeq;
     }
 
-    for (uint32 seq = minSeq; seq <= maxSeq; seq++)
+    // use uint64_t for seq to prevent overflows
+    for (uint64_t seq = minSeq; seq <= maxSeq; seq++)
     {
         auto const& envelopes = mSCP.getCurrentState(seq);
